Move method in utils
